Improve main loop
Implement the algorithm in:

https://gameprogrammingpatterns.com/game-loop.html

+++ mage/examples/fixed_camera.rs
@@ -91,7 +91,7 @@ impl Engine for GameEngine {
         Ok(())
     }
 
-    fn render(&self, _world: &mut World, _delta_time: u64) -> Result<(), MageError> {
+    fn render(&self, _world: &mut World, _delta_time: f32) -> Result<(), MageError> {
         clear(&vec![DrawingBuffer::Color, DrawingBuffer::Depth]);
         self.program.use_program();
         self.rendering_mesh.draw();
